requirement.txt
praw
python-dotenv
psycopg2-binary
transformers
torch
matplotlib
openai
nltk
gensim